Docs : 메인페이지에 있던 수정, 삭제 기능 상세페이지로 이동
- 수정, 삭제 기능 이동하면서 불필요한 코드 삭제, 필요한 코드 추가
- 로그인시 userPhotoURL 추가
- 기존에 사용하던 firebase에 users Collection 관련 부분 삭제

+++ src/components/detail/DetailComments.jsx
@@ -4,8 +4,22 @@ import { db } from '../../firebase';
 import shortid from 'shortid';
 import { styled } from 'styled-components';
 import { useParams } from 'react-router-dom';
+import { useSelector } from 'react-redux';
 
-function DetailComments({ users, post }) {
+function DetailComments({ users }) {
+  // identification
+  // [ ] detail page에서 새로고침하면 userState => 빈값되어서 오류 뜸 (이 문제는 마지막에해도..될듯)
+  // ?
+  // [ ] user확인 id 뭘로 통일할지 결정해야함 (게시물&코멘트 -삭제, 수정 시 필요)
+  const userState = useSelector((state) => state.user);
+  const userNickName = userState.displayName;
+  const userEmail = userState.email;
+  const userId = userState.uid;
+  const userImage = userState.photoURL;
+  console.log('디테일', userState);
+
+  const param = useParams();
+
   // CommentListing
   const [isLoading, setIsLoading] = useState(true);
   const [comments, setComments] = useState([]);
@@ -40,7 +54,8 @@ function DetailComments({ users, post }) {
   });
 
   // form comment
-  const postDBId = post.id;
+  // [ ] 로그인 중인 user nickName입력
+  const postDBId = param.id;
   const onChange = (e) => {
     setComment(e.target.value);
   };
@@ -59,11 +74,14 @@ function DetailComments({ users, post }) {
   };
   const onSubmit = async (e) => {
     e.preventDefault();
+    // [ ] 아래 commentWriter 관련 정보 셋중 하나 or 2개만 선택
     const newComment = {
       commentId: shortid.generate(),
       date: getToday(),
-      userId: 'userId',
-      nickName: '젤리곰',
+      userId,
+      userEmail,
+      userNickName,
+      userImage,
       postDBId,
       commentBody: comment
     };
@@ -74,12 +92,15 @@ function DetailComments({ users, post }) {
       return [...comments, newComment];
     });
     setComment('');
-    // NOTE
     setAddComment(!addComment);
   };
 
   // delete comment
-  const onDeleteCommentClick = async (id) => {
+  const onDeleteCommentClick = async (id, commentUserId) => {
+    // [ ] 댓글 작성자가 아니면 버튼 안보이게 했는데 아래 조건문도 필요하나?
+    if (commentUserId !== userId) {
+      return;
+    }
     const commentRef = doc(db, 'comments', id);
     await deleteDoc(commentRef);
 
@@ -91,23 +112,20 @@ function DetailComments({ users, post }) {
     <DetailCommentsWrapper>
       <div>comments list</div>
       <CommentUl>
+        {/* [x] 아래 user의 프로필사진 가져올 때 user확인 resource 변경됨 firestore -> auth */}
         {filteredComments.map((comment) => {
-          const commentWriter = users.filter((user) => {
-            return user.userId === comment.userId;
-          })[0];
-          let commentWriterImage = commentWriter.userImage;
-          commentWriterImage === '' && (commentWriterImage = 'https://cdn-icons-png.flaticon.com/512/552/552721.png');
-          console.log(typeof commentWriterImage);
           return (
-            <CommentLi>
+            <CommentLi key={comment.id}>
               <UserInfoArea>
-                <UserImage key={'image'} src={commentWriterImage} alt="userProfileimage" />
+                <UserImage key={'image' + comment.id} src={comment.userImage} alt="userProfileimage" />
               </UserInfoArea>
               <div>
-                <p>{comment.nickName}</p>
+                <p>{comment.userNickName}</p>
                 <p>{comment.date}</p>
                 <p>{comment.commentBody}</p>
-                <button onClick={() => onDeleteCommentClick(comment.id)}>삭제</button>
+                {userId === comment.userId && (
+                  <button onClick={() => onDeleteCommentClick(comment.id, comment.userId)}>삭제</button>
+                )}
               </div>
             </CommentLi>
           );
